mapped

+++ src/main/java/com/ticketing/ticketing_system/repositories/TicketRepository.java
@@ -1,7 +1,8 @@
 package com.ticketing.ticketing_system.repositories;
 
-import com.ticketing.ticketing_system.entities.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketRepository extends JpaRepository<Ticket, Integer> {
+import com.ticketing.ticketing_system.entities.Ticket;
+
+public interface TicketRepository extends JpaRepository<Ticket, Long> {
 }
